backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# HTTP Client
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Scheduling (for ingestion)
apscheduler==3.10.4

# LLM (optional)
ollama==0.1.6

